kunimasa/dotmaker/requirements.txt
numpy
Pillow
requests
flask
scikit-learn
